[feat] add browser polyfill to html pages
